integrate initialization functions in PS library

+++ _init/clean.ps1.tpl.ps1
@@ -1,5 +1,7 @@
 @"
-foreach (`$folder in @("bin", "obj", $ArtifactFolder, $IntermediateFolder)) {
-  Get-ChildItem -Filter `$folder -Directory -Recurse | Remove-Item -Recurse
+foreach (`$folder in @("bin", "obj", "$ArtifactFolder", "$IntermediateFolder")) {
+  if (Test-Path `$folder) {
+    Get-ChildItem -Filter `$folder -Directory -Recurse | Remove-Item -Recurse
+  }
 }
 "@
